+++ sdk/azidentity/workload_identity_test.go
@@ -1,0 +1,108 @@
+//go:build go1.18
+// +build go1.18
+
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package azidentity
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+	"time"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/internal/mock"
+)
+
+func TestWorkloadIdentityCredential(t *testing.T) {
+	tempFile := filepath.Join(t.TempDir(), "test-workload-token-file")
+	if err := os.WriteFile(tempFile, []byte(tokenValue), os.ModePerm); err != nil {
+		t.Fatalf("failed to write token file: %v", err)
+	}
+	validateReq := func(req *http.Request) bool {
+		if err := req.ParseForm(); err != nil {
+			t.Error(err)
+		}
+		if actual, ok := req.PostForm["client_assertion"]; !ok {
+			t.Error("expected a client_assertion")
+		} else if len(actual) != 1 || actual[0] != tokenValue {
+			t.Errorf(`unexpected assertion "%s"`, actual[0])
+		}
+		if actual, ok := req.PostForm["client_id"]; !ok {
+			t.Error("expected a client_id")
+		} else if len(actual) != 1 || actual[0] != fakeClientID {
+			t.Errorf(`unexpected assertion "%s"`, actual[0])
+		}
+		if actual := strings.Split(req.URL.Path, "/")[1]; actual != fakeTenantID {
+			t.Errorf(`unexpected tenant "%s"`, actual)
+		}
+		return true
+	}
+	srv, close := mock.NewServer(mock.WithTransformAllRequestsToTestServerUrl())
+	defer close()
+	srv.AppendResponse(mock.WithBody(instanceDiscoveryResponse))
+	srv.AppendResponse(mock.WithBody(tenantDiscoveryResponse))
+	srv.AppendResponse(mock.WithPredicate(validateReq), mock.WithBody(accessTokenRespSuccess))
+	srv.AppendResponse()
+	opts := WorkloadIdentityCredentialOptions{
+		ClientOptions: policy.ClientOptions{Transport: srv},
+	}
+	cred, err := NewWorkloadIdentityCredential(fakeTenantID, fakeClientID, tempFile, &opts)
+	if err != nil {
+		t.Fatal(err)
+	}
+	testGetTokenSuccess(t, cred)
+}
+
+func TestWorkloadIdentityCredential_Expiration(t *testing.T) {
+	tokenReqs := 0
+	tempFile := filepath.Join(t.TempDir(), "test-workload-token-file")
+	validateReq := func(req *http.Request) bool {
+		if err := req.ParseForm(); err != nil {
+			t.Error(err)
+		}
+		if actual, ok := req.PostForm["client_assertion"]; !ok {
+			t.Error("expected a client_assertion")
+		} else if len(actual) != 1 || actual[0] != fmt.Sprint(tokenReqs) {
+			t.Errorf(`expected assertion "%d", got "%s"`, tokenReqs, actual[0])
+		}
+		tokenReqs++
+		return true
+	}
+	srv, close := mock.NewServer(mock.WithTransformAllRequestsToTestServerUrl())
+	defer close()
+	srv.AppendResponse(mock.WithBody(instanceDiscoveryResponse))
+	srv.AppendResponse(mock.WithBody(tenantDiscoveryResponse))
+	srv.AppendResponse(mock.WithPredicate(validateReq), mock.WithBody(accessTokenRespSuccess))
+	srv.AppendResponse()
+	srv.AppendResponse(mock.WithPredicate(validateReq), mock.WithBody(accessTokenRespSuccess))
+	srv.AppendResponse()
+	opts := WorkloadIdentityCredentialOptions{
+		ClientOptions: policy.ClientOptions{Transport: srv},
+	}
+	cred, err := NewWorkloadIdentityCredential(fakeTenantID, fakeClientID, tempFile, &opts)
+	if err != nil {
+		t.Fatal(err)
+	}
+	for i := 0; i < 2; i++ {
+		// tokenReqs counts requests, and its latest value is the expected client assertion and the requested scope.
+		// Each iteration of this loop therefore sends a token request with a unique assertion.
+		s := fmt.Sprint(tokenReqs)
+		if err = os.WriteFile(tempFile, []byte(fmt.Sprint(s)), os.ModePerm); err != nil {
+			t.Fatalf("failed to write token file: %v", err)
+		}
+		if _, err = cred.GetToken(context.Background(), policy.TokenRequestOptions{Scopes: []string{s}}); err != nil {
+			t.Fatal(err)
+		}
+		cred.expires = time.Now().Add(-time.Second)
+	}
+	if tokenReqs != 2 {
+		t.Fatalf("expected 2 token requests, got %d", tokenReqs)
+	}
+}