+++ sdk/azidentity/CHANGELOG.md
@@ -3,7 +3,10 @@
 ## 1.3.0-beta.1 (Unreleased)
 
 ### Features Added
-* Added `WorkloadIdentityCredential` to support Workload Identiy Federation on Kubernetes
+* `WorkloadIdentityCredential` and `DefaultAzureCredential` support
+  Workload Identity Federation on Kubernetes. `DefaultAzureCredential`
+  support requires environment variable configuration as set by the
+  Workload Identity webhook.
   ([#15615](https://github.com/Azure/azure-sdk-for-go/issues/15615))
 
 ### Breaking Changes